Remove tmpfs at /run
Automatically mounting a tmpfs at /run causes
all sorts of interesting issues with copying stuff into the container.

In addition, there are some images (like Ubuntu, that place stuff
under /run), an example:
/run
/run/utmp
/run/mount
/run/mount/utab
/run/lock
/run/systemd
/run/systemd/was-enabled
/run/systemd/container

--
When we try to cp anything into the container under /run,
it gets confused as well, and it tries to copy to the
layer provided by the graphdriver rather than the actual
container filesystem.