Wait for the IPv6 Address to be usable before considering it done setting up
Previous to this, we would give up an IPv6 immediately to the container, but
due to the way SLAAC works, it takes a moment for the actual IPv6 address to
be usable.

This chooses a more conservative option, and waits for the address to become
usable before giving it to the user. On average, it takes about 5 seconds,
but I've seen as high as 20 seconds, when EC2 is having issues.