Add container SSHd
This commit adds an embedded ssh daemon per container. It uses an SSH
configuration which runs on port 7522 (the configuration is currently
embedded in the code).

The configuration sets up some principals, based on conventions that
our SSH CA uses to specify what account / app / instance a user has
access to.

The way it actually does this is by creating a Docker container, with a
volume that has a relocatable SSH daemon installed within it. This
Dockerfile is in the project. Once this volume is mounted within the
container, the executor sets up the config file, the principals allowed,
and copies the allowed CA list from the host to the container.

After this is done, it starts up a templated systemd unit, called
"titus-sshd", which performs a custom nsenter, and sets the apparmor
profile of the SSH'd users to the same as the container.