make it easier to override the nflx env file
all other nflx env files provided by nflx base images follow the
convention of prefixing files in /etc/nflx/base-environment.d/
with a number in the NNN format. This will make it easier to
override variables defined in the file supplied by the titus
executor, and ensure that overrides are loaded after the titus
file (for the common use case of loading files in the alphabetical
order).