Remove Logging to Journald only under systemd
This makes it so that the executor doesn't rely on CGO. It needed
cgo before to figure out if it was running as a systemd service
or not. This makes it assume it will always run as a systemd service
unless otherwise stated by console flags, rather than relying on
environment detection.