Fix erroneous read from closed pipe
Calling .Wait() makes it so that the Stderr / Stdout pipe are closed
when the command finishes. This makes it so that we're unable to read
the JSON that comes off of the command. We should wait to call .Wait()
until *after* we've finished reading the JSON, or given up on reading
the JSON.