Only allow overrides which point to same registry.
Code review uncovered the possibility of launching a DoS attack on
titus by pointing the container attribute to a different registry
that straves the host resources.

This change makes it so that the service mesh container no longer
specifies the registry. Only the container name and digest will be
specified in the property and the conatiner attribute.