Consider subnets not found error as non-fatal, NotFound
This error can happen in one of three conditions:
1. The agent's account / task's account / az being asked for has
   no subnets configured for it. This would be either caused
   by a misconfiguration on the Titus team's part, or the
   user's part
2. The data we fetch from AWS around subnets is wrong,
   and we've lost track of the subnets which we launch
   in.
3. Misconfiguration - Either the control plane, or the
   agent have misconfigured the default account / AZ / subnets
   for tasks.

Recently, we've seen an uptick in cases where the inputs
-Agent account
-Task requested account [optional]
-Task requested subnets [optional]

Have not aligned. We haven't seen grave misconfiguration of
our services, or the EC2 APIs return bad data lately.

We can alert on notfound as a % of total launches at a
much higher percentage if we want to keep track of
these errors still.