getGCableENIs now checks context, and exits when it's done
getGCableENIs previously would iterate over all ENIs, and then and
write them to a channel. This was a buffered channel that had a
depth of 10,000. Simultaneously, there was a context that was used
by the describeENIsFoGCWorker that read off this channel to
abort reading off the channel. If the context was cancelled,
the describeENIsFoGCWorker would stop reading off of the
(describe) channel.

This was previously never a problem because the size of the buffer
was 10,000, and we never had more than 10,000 ENIs. Now that we have
more than 10,000 ENIs, it means that this describe loop can stall
out.

I wish that Go channels were more like Erlang, and that they were
unbounded, and if they made memory run out, they would panic,
or you could get an error on write. That seems far more sane.