Fix environment variable resolution
I introduced a bug that allowed people to override the environment of Titus containers.
There are certain environment variables that are special, and sould not be able
to be overridden. These are set by the Titus executor.

There is one special one where the control plane can set it --
EC2_OWNER_ID.