This scaffolding is for my intermediate project: Getting sshd working on kubelet again.
To do this, I want to re-use as much as the titus-executor code as I can.
Previously I didn't do that, and it broke, and and it drifted in behavior
(see https://github.com/Netflix/titus-executor/pull/487).

So my new goal is to implement ssh in such a way that,
as we improve/change ssh on VK, it automatically applies to RK,
and we don't have to continuously play catchup.

To do that, I propose this oci hook binary to use in
https://github.com/Netflix-Skunkworks/titus-kubelet/pull/39.
A single hook for all the things, to reduce the burden of making
different binaries for all the things, and in theory no additional
changes to titus-containerd or any other hook configuration to implement new stuff.