Added the Titus Seccomp Agent (tsa)
This new system sidecar (currenly vk-only) intercepts
syscalls on behalf of the container.

The first style of interception is to enable perf-related
syscalls (perf_event_open, bpf) where they were otherwise
not allowed (and wouldn't function even if they were allowed).

This feature is behind a container param:

    titusParameter.agent.seccompAgentEnabledForPerfSyscalls

That instructs the agent to setup the:

* Seccomp policy with perf/bpf allowed
* Env variables for tini to consume to relay the seccomp notify socket
* To spin up the tsa agent itself under systemd

The reason this is behind a perf-specific bool is that we intend
to expand the scope of this agent to someday include network
syscalls too, but not all containers are going to want both features.
Having different bools will allow us to enable new interceptions without
affecting containers that have opted in for only one kind of behavior.