Add second IAM proxy for log profile.
We need to provide credentials to the log upload profile in the network
namespace of the pod. We add an env variable to the imds proxy to add a
second assume role proxy if started by the imds-cni hooks.

These credentials are available at /latest/meta-data/iam/logging/...

Additionally added a credential provider to the uploader package for
loading them.