Fix the imds local_ipv4 env variable
The full ip object includes the CIDR, so things like:

    192.0.2.1/18

Were being written down, latter to be consumed by net.ParseIP,
which doesn't understand that, so it returns nil.

This fix is to always write down on the IP portion (not the mask), so
that it can be parsed correctly later.