Fix accountId in kubelet IMDS instance-identity document
The NETFLIX_ACCOUNT_ID is set by TJC, we can't use whatever it is
set to on the agent (could be any crazy account ID).

The IMDS uses this when it returns data for

    curl http://169.254.169.254/latest/dynamic/instance-identity/document

Previously this was unset, causing `accountId` to be the empty string.
With this change it will show the account ID set to whatever is set
in the main container's environment.