WIP: titus-storage system service with ebs support
This adds a new system sidecar: `titus-storage`.
The goal of this sidecar is to mount all external storage,
ebs to start, and nfs later (or if you want I can do nfs first?).

I think there could be few nice things about doing it this way:

* Long term could be externalized into a "true" external sidecar
* Way easier to troubleshoot / debug (it is a systemd unit, not an
  internal titus-executor thing)

Some downsides:

* Being external to the titus-executor codebase, most of our message
  passing comes via environment variables (Hard to do N ebs mounts?)
* I don't know what to think of having multiple mount binaries, but
  same deal, could be externalized to other teams some day
* This system service *must* wait, otherwise there isn't a good
  way for it to "stop" and detach.
* All attach/detach is happening client side, not centrally orchistrated
  like the aws-ebs-csi driver.

This POC is in bash because a lot of bash code existing already.
For production I would rewrite it in golang, with real error codes
and not having to "tail" to wait for a pid.