Added titus-mount-block-device command
This new command is analogous to titus-mount (which does NFS only)
but it operates on any block device.

This is in preparation to support EBS block devices.

Some day we might merge these code bases a bit more, and since
I'm putting them in the same folder, there is room.
But for this experimental EBS support, don't want a chance
of breaking nfs.

The nfs mount does have different requirements, including
dns lookups, a different ordering between when to switch
the namespace, etc.

In practice, I don't know if there are any other "types"
of things to mount? Maybe real CEPH or other things, but
we can wait to see if we need to merge these codebases
or do something different.