Added the titus-storage service, with EBS support
This is the new titus-storage system service.
Unlike previous storage options (nfs), this can
run as a system service, which should make it easier to debug
and iterate on. Eventually I would like for it to handle
other storage needs, perhaps nfs too.

It is pretty conservative, and will refuse to do anything scary.

Probably my biggest fear is around leaving orphaned attached volumes
that need to be manually cleaned up. I don't know how often this will
happen, but I think we'll need a reaper.