Make the waitForTitusIsolate function properly timeout
The select call is pseudo-random, you can't control
which messages it selects.

In prod, we had instances where this function was never
timing out, expending the entire task budget time
waiting for titus isolate.

I believe this was because the ticker spammed the channel,
and the done message never got in on time.

This PR makes it so we give Done a fighting chance, based off
of https://stackoverflow.com/questions/46200343/force-priority-of-go-select-statement/46202533#46202533

Playground repro:
https://play.golang.org/p/soqUFWhZ7sj

And a solution:
https://play.golang.org/p/Dw96oMVZJst