Make titus-mount-block-device use fd instead of pid
In general fds are more robust than PIDs, but with
titus, it isn't super easy to know what this PID
of your container actually is.

But we do know a well-known path. From that path,
with some C trickery, we can get all the namespaces
we would have needed otherwise and gotten from a pidfd.