Prevent invalid environment variables from slipping into profile.sh
We don't strictly enforce the env variable strings at our
outer API layers. I think that is "kinda" ok. Maybe
some day we'll be stricter.

But deep down where we are actually deploying a shell-compatible
file, then we can not put in shell-incompatible variables.

But, for discovery, I put the invalid ones at the bottom and comment
them out.