Enable titus-seccomp-agent for sshd
titus-seccomp-agent is used to intercept syscalls
for the main process of a container.

It would be confusing if, when you ssh'd to that container
those same syscalls didn't work!

This change makes it so sshd-spawned processes behave the same.

Maybe in the future we'll have other nsentered processes
also utilize TSA.

This makes the seccomp notification slighty more robust, instead
of needing full sidecar ordering.