Make titus-seccomp-agent die when a task finishes
Unlike many other system sidecars we have, TSA doesn't
inject itself into the pid namespace of a container (yet)?
and doesn't die with the rest of the container "magically".

To make sure it terminates when titus-executor does, preventing
tsa's from hanging around, I make it PartOf titus-executor.

Also, I added a --titus-task-id= command line flag to TSA.
This is kinda a hack, because it is ignored by tsa itself,
but makes it super easy to know which TSA goes to which process
by simply running ps.