Selectively install NO_NEW_PRIVS when adding a seccomp_filter
Including NO_NEW_PRIVs is mostly incompatible with apparmor
profile-switching, which is cumbersome for ssh, but the bigger
showstoper is that sudo stops working.

(nfsuper) ~ $ sudo bash
sudo: effective uid is not 0, is /usr/bin/sudo on a file system with the 'nosuid' option set or an NFS file system without root privileges?

This is similar to what we saw in real kubelet, which had
NO_NEW_PRIVS on all the time.

But that is OK, we can have our cake and eat it too, but only
doing NO_NEW_PRIVS when we *don't* have CAP_SYSADMIN.

This means SSH can still be root, and get the seccomp_filter.

I don't know what the implications are of other workloads
at Netflix. I bet there are some workloads that are trying
to use setuid binaries (sudo), which will not work after this
is in place and they start using TSA.