Always run stop sequence on failed titus-storage starts
Normally we get a 'stop' command when a container dies
or gets killed.

But if the container never even starts up completely,
titus-storage will never get a 'stop' command,
because we the `titus-container` target was never reached
in the first place!

Running the 'stop' sequence if there is any error ensures
we do any cleanup we need here.