Remove custom sslTitusAgentCA logic from titus-vpc-service
This sslTitusAgentCA logic was a mistake, I didn't really
understand how mutualtls worked on grpc, and thought
we needed to do additional verification on the certificates
in this auth function.

This auth function used the 'web.crt', which happened to have
the intermediate CA, so it "worked". However, when that
intermediate CA rolled, there was mistmatch between the
client and server, so this verification function failed.

It failed, *even though* the golang tls stuff was doing its
job by verifying that the 'web.crt' was *actually valid* by
traversing all the way up to the normal SSL CA stuff.

This PR removes my not-needed verfification in favor of
the normal stuff. A different PR (#427) will fix the todo
and add *authz* to complement the mtls authn.