Adjust titus-ssh to be compatible with scratch containers
This allows titus-ssh to work on minimal containers
(tested with scratch & alpine).

And in general, will allow ssh to work on any wacky
container our users can come up with, and still
have "the titus experience" where they can ssh
to it and get at least some sort of shell.