Implemented first MVP of multi-container user workloads (pods)
This is an MVP of multi-container workloads on Titus (pods).
It uses the "Country & Western" approach, which treats the
first container as the "main" one, which holds all the namespaces.

The lifecycle of a pod in this MVP is:

1. Start up, but pause (via tini) the main container
2. Start up the Titus system services (Country style)
3. Start up the Other containers (Western style)
4. Unpause the main container.

The main advantage to this design is that when there is only
1 container, step 3 is a noop. This provides backwards
compatibility with our existing workloads while we iterate
on the nuances and functionality of step 3 (ordering, volumes,
healthchecks, logs, etc)

----

With this change, the `Runner` struct in the executor
now holds the whole pod object. This allows the container
creation to have all the information it needs to properly
launch (Run) the main workload *and* all the sidecars.

I think eventually we'll need to tease out the `container`
object from the `Runner`, separating the settings that
are for the whole *task* (network stuff, global params),
and then have an *array* of container objects, instead
of the singleton, but I would like to do that refactor
on another pass, for when we need more direct access
to the extra container objects.