Ensure tini is always available to every titus container
I want to make sure that we can get "the titus experience",
especially when it comes to multi-container workloads (pods),
*even* on a mac laptop.

This means making sure we always have tini. Once we can safely
assume we always have tini, then we can remove remove codepaths
which cannot assume it to be there.

To do this, I bind mount tini from the host, instead of depending
on --init being set on the docker daemon. This works even if it
*is* set.