Added beta ephemeral storage support via titus-storage for GPU tasks
This adds Beta support for having a /ephemeral directory
mounted inside the container that represents scratch disk.

This is most often requested for GPU-based training workloads
(also the only place were we have ephemeral disks).

This is a simple implementation at first, with no regards
for quota or protection, just minimal isolation. This
feature will be considered "best effort" for our users.