RFC: Use static busybox and static openssh binaries for titus-sshd
This is a radical departure from the way we build titus-sshd
today, but is faster and hopefully more sustainable?

It uses the sshd binary from the minos distro, which has a
library of static binaries:

https://github.com/minos-org/morpheus-ports/tree/master/openssh

The downside is that this is a step backwards in versions
6.1 compared to what we currently run, 7.7. And the build is
more opaque.

On the upside, the image is much smaller, because things are
statically linked instead of just relocated.