Added misc error improvements and macos docker support
Running the standalone tests on osx is pretty close
to working, the only thing missing was turning
of tini bump priority.

However, to get there, I had to debug pretty hard because
the only thing I had to go off of was "operation not permitted",
so I left in additional error handling code to make
this type of debug easier for the next person.

I also just made the tini thing non-fatal, as I don't think
it should get in the way of a container launch.