Added titus-network mode
This new network-mode is designed to be a "source of truth"
for what kind of networking to setup on a container.

Specifically I want some way to know for sure if TSA and VPCservice
are supposed to setup a ipv4 transition mechanism.

OR, in the future case where we know we want ipv6 only, vpcservice
needs to know that and forgo certain things.

Likewise, titus-inject-metadata will probably have to behave
differently. And also I even expect user applications to take advantage
of the `NETFLIX_NETWORK_MODE=IPV6_ONLY` style hints, so that they
don't have to scrape ifconfig or other weirdness, and can adjust
their behavior appropriately.

This just plumbs the setting through, no additional code is in place
yet to actually act on it.