Refactor: Launch all (main and other sidecars) in together
This is 'just' a refactor to make it so I have a
single function responsible for starting all
containers, the main one (via) tini included.

This will allow us to (in future PRs) be able
to do full ordering amongst all defined containers,
where "main" is just a special case (launch via tini
instead of launching via `ContainerStart`).

My end goal here is to have a function that can take
the whole pod, and report back the ordering required,
and then my `startAllUserContainers` function will
be able to start things in the right order (including the
main one).