WIP: Added basic ordering (class type) in multi-container workloads
This is the first part of class-based sidecar ordering in
Titus. It does not do "dependsOn" yet, only does two-layer
ordering between platform-defined and user-defined containers.

Currently just starts all platform-containers in parallel,
and then user-defined ones in parallel. dependsOn will be complex
and added in the next pass of code.

For this one, I wanted to check to make sure I was going in the
right direction.