Misc logging and status updates
When trying to debug why a container didn't start
correctly, I took it as an opportunity to do a
refresh of all the (semi) cryptic `creating` and
`got msg` log messages.

My main philosophy here is to remove the "cool story"
logs, and hopefully make it easier for users and operators
to know the "why" for what is happening.