Added NETFLIX_IPV6 vars
A Netflix we (apparently) do not use the EC2_IPV6 variables.
This is because we don't trust them, because they don't come
from v6 assigner, which come post-launch. Instead they use
the Netflix-specific NETFLIX_IPV6* vars.

In order for application to smoothly transition to Titus, this
PR adds the same variables as well.