Fix the titus-sshd docker build for 2021
I think the gentoo approach is OK. It does give us
the benefit of being able to precisely control how
sshd is built, *and* keep all the features of a modern
ssh daemon (port forwarding, ssh CA auth, etc).

The downside of course is... gentoo and build times.

But this does work, and maybe it is just worth it to
fight the build once every couple of years.