Re-enable sshd on scratch containers using a startup script
This re-enables the run-titus-sshd script which makes it
so we can ssh to scratch containers.

This was previously reverted because we were blocked on
updating the sshd image. Now that it is updated and deployed,
we can safely do this bump and enable sshd on scratch containers.

(or an container that doesn't have the things sshd needs, like
the nobody user, etc)