Move around security group checking logic
This does a bunch of refactoring related to where security groups logic
is handled. Because the transition entry will be its own entry in
assignments, we need to change where the logic is, if we're to try
to do work across two discrete transactions (i.e. generate two
assignments, and populate them vs. trying to populate two assignments
in the same code). By shifting the logic into the DB of doing
verification at "commit" time, we simultaneously insert two
assignments in the same transaction without problem.