Allow users to specific version of titus system sidecars
This allows users or titus developers to specify the version
of a system sidecar (or set to be empty!).

This is useful for platform partners who want to unset sidecars,
allowing them to opt out if they want.

It also allows for sidecar developers to better iterate on them,
because they can override which version they want to use, so they
can canary a sidecar before anyone else sees it.

If shipit, I'll setup our functional tests to be able to set these
overrides, so that we can run our tests against sidecars that
are not even really deployed yet.

This override kinda already existed for servicemesh, but this generalizes
it to any sidecar.