Added an FSlock on titus-storage
This adds a fs lock on titus-storage, to avoid race condition when
multiple mounts/attaches are happening on an agent.

This was recommened in the beginning, so here it is after seeing the
race conditions in the wild.

I tried really hard to add sleep and debugs to see this in action,
but for some reason I never see `/tmp/titus-storage` on my agent,
even though I'm certain the right code is running (prints).
Is there something I'm missing?