Statically wire up the entire IMDS, and handle link watch
There were a bunch of race conditions in the IMDS due to how certain
behaviours were being handled in the Linux kernel (for example, veth
bringup was / is async). This in turn, makes sure that's all okay.

Beyond that, it does static programming of the neighbor table,
so we don't have to take any chances on neighbor discovery
"acting weird".