Added base environment support to country & western pods
You could argue that some (not all) of this could be injected in the webhook.
One downside so that approach is that it would be split: we would
*still* have to have this `Env()` function in the executor to actually
compute the whole env (docker image vars, vpctool vars, etc). The other
downside is that populating the entire env in the pod would make our pod objects
*even* bigger, duplicating the arrays N times for N containers.

Some of these variables are not known to the webhook until the last
second, and they are not "container-specific" you might say.

When a user specifies an env variable for the whole job, they probably
expect it to show up on each container. This PR implements that.

It also starts a new variable, `TITUS_CONTAINER_NAME`, which we did have
on kubelet, so I'm bringing it back.