Added new Pod task identity endpoint to the titus-imds
This new taskPodIdentity endpoint is very similar to the
existing taskIdentity endpoint, except instead of providing
containerinfo, it provides a full pod. This will provide
for attesting task identity based on containers (plural)
instead of containers.

Additionally, by providing the whole pod, we can just let
Security (metatron) do whatever they want based on the pod
object, no changes needed on our side.

The downside is that the pod object isn't quite as structured
when it comes to the metadata that metatron wants. Any consumer
of this pod will have to parse the annotation strings.

Another potential security issue here is that we are now exposing
the pod object to the pod processes! It may be a slippery slope
if we make this a public api to other consumers.

Another potential issue is that this pod object is loaded on titus-imds
start, and frozen. We don't reload the json, even though things like
podstatus do get updated in that file. They may be confusing?

Until this is finalized, I made the loading of this object not-fatal
to the imds. If there is any error, the show will go on.