Refactor: Promote user and platform containers as real attributes of the runtime struct
In some sense, this is much cleaner than constantly parsing the pod object
all the time.

It does solidify the difference between user/platform containers,
hopefully that is a good thing.

Someday, the bit `Container` object can be broken down in to the job stuff
and the `UserContainers` can have all user containers, not just the
extra ones!

The reason for this refactor is to make it easier to store more methods
and attributes into the ExtraContainer struct to make it easy to do
bookkeeping with them. You can see the first thing that becomes
easier is tracking the container id for each one. I no longer need
and extra map hanging around just for that, and I can iterate over
the real objects instead of their names. Good stuff.