Added v0 of container healthchecks
Surprising to me, users can add healthchecks to their docker *images*
today! We just don't do anything except report it on the updates channel.

This PR adds support for accepting liveness probes from the pod spec
and converting them to use docker healthchecks on the backend.

This is a feature we get "for free" because of the C&W design (using
docker natively as the executor).

This currently doesn't help the user very much, because I don't have
the wiring in place to see health events. I'm waiting on my big
refactor PR to plumb that through.

However, soon we'll need to address a big elephant, which is "what"
to do with this data. Surprising to me (TIL), healthchecks don't
have anything to do with restart policy!

We will have to implement our own connector, like kubelet does
to kill the container once it goes unhealthy, and let the restart
policy take effect.
https://github.com/moby/moby/issues/28400

But that can be in v1 (this is v0).