Added Pod status container statuses for multiple containers
Ever inching towards the multi-container world.

This plubming makes it so we can track the status
of the other non-main containers.

I'm not sending additional pod updates, however.
I'm concerned about overwhelming the API server.

But, when it is updated, I should be able to see the podstatus
via kubectl and actually see how my sidecars are doing.

Next steps after this are:
1. Add docker event watchers to actually update the status after `start`.
   Optionally killing and restarting containers

2. Update the Titus API so that we can actually get the containerstatus's
   all the way back to the user (so one doesn't have to use kubectl)