Only make metatron creds on a pod object if we have some
Previously, we would create a metatronCreds entry in our cInfo
object for pods, *even* if we didn't have creds to work with.

Even though the struct instance had null fields, we still had a struct.

This makes it so the struct pointer stays nil if we don't have
metatron creds in the first place.