Send the effective networkMode back to the control-plane
We only get 1 ipv4 and 1 ipv6 to send back to the control-plane
in the form of the pod status.

In the IPv6Only+transition mode, that ipv4 represents the transition ip,
and should not be considered the "normal" ip for the pod.

For this reason, it is helpful to send the effective network-mode
back to the control-plane with the rest of the network configuration
updates, so that it can make better decisions about what data (ip addresses)
it is getting.

Pairs with
https://github.com/Netflix/titus-control-plane/pull/1086