Sprinkle in better messages and errors on kill/lost
Right now on task Kill, the "reason" is always the same
message as the previous state, which often "main container running".

This PR makes it so that on task killed/lost, we set the message
to be something more appropriate, *and* we append the error messages
if we have any.