Refactor the statusMessageChan update stuff
The statusMessageChan is a channel that transports updates from
a the docker runtime's health to the runner.

As we move towards multiple containers, we are going to need
to get updates for different containers, any of which can oom, die,
etc.

This refactor prepares the statusMessageChan machinery for multiple
containers, particularly away the closing of the channel as a method
to trigger titusdriver.Lost (should never happen).