RFC: TitusMainContainerVolume mount support
I expect a common pattern in Titus will be to
"Share this volume between the main container".

This would power things like Fuse mounts from a fuse sidecar,
or just sharing /mydata to things like nginx or whatever.

Our experience when working with this kind of pattern in
stock kubelet is kinda painful, becuase it means making
and EmptyDir, making an init container to copy files over
from the main one to that dir, and then sharing that EmptyDir
volume with another container.

But we control the runtime, we can make the common thing easy,
and I bet 90% of the kinds of volumes we will want will be like this.