Undo Prefix Delegation
This commit does two things:
 * Removes the logic to always assign a prefix to each branch ENI
 * Uses an alternative method to assign IPv6 addresses

The way the logic works for assigning IPv6 addresses:
* We introduce a new structure in the database, "worker_id". At
  the startup of each instance of the titusvpcservice, it gets
  a worker ID. There are only 4095 worker IDs available
* Every time we assign an IPv6 address to a container, we use a
  a per-subnet incrementing counter
* For a given subnet, we start the counter (when the first IP
  is assigned), at a random number.

The risk here is if there is a random number collision, and the
subnet is densely populated. If this happens, we may be able to
alleviate the situation by restarting the service. One other
choice is to hoist this to the database, but that involves /
requires that we do at least one more DB write per assign.