Added inspect info for all containers, not just the main one
For properly configuring extra containers for entrypoint/args, systemd
and tini, I need to inspect them, just like we docker image inspect
the main container.

Luckily we already have the function to do this, it just needed a little
bit of tweaking.

It is unfortunate that the docker image provided from k8s is not in
collapsed form. Luckily we already have functions to deal with that!