Make sure that all containers in a pod have the same security settings
Previously, I was not setting the dockerconfig setting for security
options.

This PR reuses our existing function to set these on the main container,
except now it applies to *all* containers in a pod, ensuring they all
share the same seccomp policy and apparmor profile.

TINI_HANDOFF was taken out because it isn't really a security setting, and
I didn't want it applying to extra containers (They are not running systemd)