Added a command to switch between multiple containers in a pod
Not as advanced as the "run one sshd per container", but is cheap.

I don't think we should consider this a stable interface, but a
good enough solution for sidecar authors to troubleshoot their stuff.

If shipit, I can add new links (once the titus api understand podstatus output)
for each container, which a one-click link that copies

    ssh $task -- switch-container $container

To make it easy